Move views from activity to separate files
